Move trigger map related models to trigger_map.go and trigger_map_item.go
