+++ models/trigger_map.go
@@ -1,0 +1,67 @@
+package models
+
+import "fmt"
+
+// TriggerMapModel ...
+type TriggerMapModel []TriggerMapItemModel
+
+// Validate ...
+func (triggerMap TriggerMapModel) Validate() error {
+	for _, item := range triggerMap {
+		if err := item.Validate(); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func checkDuplicatedTriggerMapItems(triggerMap TriggerMapModel) error {
+	triggeTypeItemMap := map[string][]TriggerMapItemModel{}
+
+	for _, triggerItem := range triggerMap {
+		if triggerItem.Pattern == "" {
+			triggerType, err := triggerEventType(triggerItem.PushBranch, triggerItem.PullRequestSourceBranch, triggerItem.PullRequestTargetBranch, triggerItem.Tag)
+			if err != nil {
+				return fmt.Errorf("trigger map item (%v) validate failed, error: %s", triggerItem, err)
+			}
+
+			triggerItems := triggeTypeItemMap[string(triggerType)]
+
+			for _, item := range triggerItems {
+				switch triggerType {
+				case TriggerEventTypeCodePush:
+					if triggerItem.PushBranch == item.PushBranch {
+						return fmt.Errorf("duplicated trigger item found (%s)", triggerItem.String(false))
+					}
+				case TriggerEventTypePullRequest:
+					if triggerItem.PullRequestSourceBranch == item.PullRequestSourceBranch &&
+						triggerItem.PullRequestTargetBranch == item.PullRequestTargetBranch {
+						return fmt.Errorf("duplicated trigger item found (%s)", triggerItem.String(false))
+					}
+				case TriggerEventTypeTag:
+					if triggerItem.Tag == item.Tag {
+						return fmt.Errorf("duplicated trigger item found (%s)", triggerItem.String(false))
+					}
+				}
+			}
+
+			triggerItems = append(triggerItems, triggerItem)
+			triggeTypeItemMap[string(triggerType)] = triggerItems
+		} else if triggerItem.Pattern != "" {
+			triggerItems := triggeTypeItemMap["deprecated"]
+
+			for _, item := range triggerItems {
+				if triggerItem.Pattern == item.Pattern &&
+					triggerItem.IsPullRequestAllowed == item.IsPullRequestAllowed {
+					return fmt.Errorf("duplicated trigger item found (%s)", triggerItem.String(false))
+				}
+			}
+
+			triggerItems = append(triggerItems, triggerItem)
+			triggeTypeItemMap["deprecated"] = triggerItems
+		}
+	}
+
+	return nil
+}
